Fixed 3rd pane
Went ahead and created scaling to find min and max for the current
selected files within the selected box.
Then had to get the scaling to work for drawing the horizontal axis
properly so that they data would be shown fully.
Only thing left would be to fix the selection so when you unclick the
box and the selection goes away that the bottom panel shows all data
again.