Added Dates to long bar
Used new function change date to turn text date into month and year
string to be displayed. Initialized upper and lower dates to be the
whole list if no box had been selected.
Only error found to date is if you drag the select box then resize the
window the box does not move relative to the screen and can get a null
pointer.